Move S3 methods from mirroredModel to new s3 file
